refactor: consolidate validation handler into main DAG handler file
- Move ValidateDAG handler method from separate file to dag_handler.go
- Move validation-related types (ValidateRequest, ValidationResultPresenter, etc.) to main handler file
- Delete separate dag_validation_handler.go file for better code organization
- Maintain all existing functionality and test coverage
- Improve code organization by grouping all DAG endpoints in single file

This consolidation improves maintainability by keeping all DAG-related HTTP handlers together rather than split across multiple files.
